+++ posthog-node/test/extensions/sentry-integration.spec.ts
@@ -1,4 +1,4 @@
-import { PostHog } from '../../index'
+import { PostHog } from '../../runtimes/index.node'
 import { PostHogSentryIntegration } from '../../src/extensions/sentry-integration'
 jest.mock('../../src/fetch')
 import fetch from '../../src/fetch'
